requirements.txt
fastapi
uvicorn[standard]
jinja2
sqlalchemy
pymysql
jinja2
aiofiles
request